classes in java

+++ Foundations of Object Oriented Programming/JavaBasics/Javabasics/src/App.java
@@ -105,7 +105,7 @@ public class App {
 
         // Explicit Typecasting - When you intentionally have to typecast a value
         int x1 = 20;
-        short x2 = x1;
+        short x2 = (short) x1;
 
         // Type mismatch: cannot convert from int to short
         System.out.println(x2);
